requirement.txt
streamlit
pandas
numpy
torch
transformers
sentence-transformers
chromadb
groq
python-dotenv
python-dateutil
graphviz
